+++ aiosmtpd/docs/NEWS.rst
@@ -7,6 +7,7 @@
 
 Added
 -----
+* Test for ``SMTP.__init__`` behavior after taking out code that edits TLS Context
 * Implement mechanism to limit the number of commands sent (Closes #145)
 
 Fixed/Improved
@@ -15,6 +16,7 @@ Fixed/Improved
 * Implement & enforce line-length-limit, thus becoming Compliant with RFC 5321 § 4.5.3.1.6
 * Delay all SMTP Status Code replies during ``DATA`` phase until the phase termination (Closes #9)
 * Now catches ``Controller.factory()`` failure during ``Controller.start()`` (Closes #212)
+* :class:`SMTP` no longer edits user-supplied SSL Context (closes #191)
 
 
 1.2.2 (2020-11-08)